+++ CHANGES.md
@@ -4,6 +4,7 @@
 * Support for Context-based urls, thanks to [Wout Neirynck](https://github.com/wneirynck).
 * *BREAKING*: top-level swagger options are now in align to the compojure-api:
   * Fixes [#22](https://github.com/metosin/kekkonen/issues/22)
+  * By default, `api`s don't bind swagger-spec & swagger-ui, use `:spec` & `:ui` options
 
 ### Old